separate base simulators into different files
