+++ test/unit_tests/braket/default_simulator/test_simulator.py
@@ -16,7 +16,7 @@ import pytest
 
 from braket.default_simulator import observables
 from braket.default_simulator.result_types import DensityMatrix, Expectation, Probability, Variance
-from braket.default_simulator.simulator import BaseLocalJaqcdSimulator
+from braket.default_simulator.jaqcd_simulator import BaseLocalJaqcdSimulator
 
 
 @pytest.mark.parametrize(
